fix(debuginfo): Remove depth limits from function parsing (#1065)
We will now handle function parsing in both symcache convertor and dwarf DIE parsing iteratively, so there is no longer a need for bounding these by function depth.

+++ symbolic-symcache/src/writer.rs
@@ -73,7 +73,7 @@ pub struct SymCacheConverter<'a> {
 struct InProgressFunction<'a> {
     function: &'a Function<'a>,
     base_index: u32,
-    depth: u16,
+    depth: u32,
     call_locations: Rc<[(u32, u32)]>,
 }
 
@@ -487,7 +487,7 @@ impl<'a> SymCacheConverter<'a> {
         tr: &dyn TypeResolver,
         function: &Function<'_>,
         base_idx: u32,
-        fn_depth: u16,
+        fn_depth: u32,
     ) {
         if !self.collect_variables || fn_depth > u8::MAX.into() {
             return;
